Module project2::vmexit::pio

Port-mapped IO vmexit controller.

Port-mapped I/O is a method of communication between external devices and the CPU, where each device is assigned a unique port address that is used as an operand in the in/out instruction family. This allows the CPU to directly read or write data to the devices. There are different variations of the in/out instructions, depending on the type and size of the operands used.

In virtual machine management, the VMM can choose to allow or trap port-mapped I/O instructions executed by the guest. In this specific project, all such instructions are trapped to the host, giving the VMM control over their behavior.

Tasks

In project2, You need to decode the 18 in/out instructions. Initialize the IoInstruction instance with the corresponding opcode and registers. The controller will then use this information to forward the request to the appropriate handler. When handling Outsw_DX_m8, Outsw_DX_m16 and Outsd_DX_m32, it is necessary to copy the memory contents by translating guest virtual address to host virtual address. You can translate the guest address to the host address by Probe.

Structs

Controller

Pio vmexit controller.

IoInstruction

The decoded information of io instruction.

Enums

Direction

Direction and the Value of the instruction

Prefix

Possible prefix of io instruction.

Traits

PioHandler

Trait that represent handlers for port-mapped devices.